Data availability

Returns the data availability. Response: DataAvailabilityID - ID of the resource as found in the database DataAvailabilityName - explanation for the data availability IsAvailable - Y for data is available, N for data is unavailable, and NULL for unknown data

Usage

cvd_data_availability(
  time_period_id = 1,
  system_level_id = 1,
  indicator_id,
  metric_category_type_id
)

Arguments

time_period_id

integer - the time period to return data for (compulsory)

system_level_id

integer - the system level to return data for (compulsory)

indicator_id

integer - the indicator to return data for (optional)

metric_category_type_id

integer - the metric category to return data for (optional)

Value

tibble of data

Details

CVD Prevent API documentation: Data availability

See also

Examples

test <- cvd_data_availability(time_period_id = 3, system_level_id = 5)