Indicators

Returns all indicators and data for a given time period and area. Also returns time series data for all time periods available. If tags are specified, only indicators which have one of the specified tags will be returned.

Usage

cvd_indicator(time_period_id = 1, area_id = 1, tag_id)

Arguments

time_period_id

integer - time period to return data for (compulsory)

area_id

integer - area to return data for (compulsory)

tag_id

numeric vector - allows filtering indicators by one or more tags (optional, array)

Value

list of tibbles

Details

CVD Prevent API documentation: Indicator

See also

Examples

test <- cvd_indicator()
test <- cvd_indicator(time_period_id = 17, area_id = 55, tag_id = c(12, 13))