Indicator data

Returns all metric data for a specified indicator. Data will include values for both selected area, and organisation at National Level (usually England).

Usage

cvd_indicator_data(time_period_id, area_id, indicator_id)

Arguments

time_period_id

integer - time period for which to return data for (compulsory)

area_id

integer - area for which to return data for (compulsory)

indicator_id

integer - indicator for which to return data (compulsory)

Value

Tibble of details for the indicator in the area and a national comparison

Details

CVD Prevent API documentation: Indicator data

See also

Examples

# Look at 'AF: treatment with anticoagulants' (indicator ID 7) in time
# period 17 for  'Leicester Central PCN' (area_id 701) focussed on metrics
# by gender:
cvd_indicator_data(time_period_id = 17, indicator_id = 7, area_id = 701) |>
  dplyr::glimpse()
#> List of 3
#>  $ indicator_metrics: tibble [22 × 17] (S3: tbl_df/tbl/data.frame)
#>   ..$ AxisCharacter         : chr [1:22] "%" "%" "%" "%" ...
#>   ..$ FormatDisplayName     : chr [1:22] "Proportion %" "Proportion %" "Proportion %" "Proportion %" ...
#>   ..$ IndicatorCode         : chr [1:22] "CVDP002AF" "CVDP002AF" "CVDP002AF" "CVDP002AF" ...
#>   ..$ IndicatorFormatID     : int [1:22] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ IndicatorID           : int [1:22] 7 7 7 7 7 7 7 7 7 7 ...
#>   ..$ IndicatorName         : chr [1:22] "Patients with GP recorded atrial fibrillation and with a CHADS2 or CHA2DS2-VASc score of 2 or more, who are cur"| __truncated__ "Patients with GP recorded atrial fibrillation and with a CHADS2 or CHA2DS2-VASc score of 2 or more, who are cur"| __truncated__ "Patients with GP recorded atrial fibrillation and with a CHADS2 or CHA2DS2-VASc score of 2 or more, who are cur"| __truncated__ "Patients with GP recorded atrial fibrillation and with a CHADS2 or CHA2DS2-VASc score of 2 or more, who are cur"| __truncated__ ...
#>   ..$ IndicatorOrder        : int [1:22] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ IndicatorShortName    : chr [1:22] "AF: Treated with anticoagulants (CVDP002AF)" "AF: Treated with anticoagulants (CVDP002AF)" "AF: Treated with anticoagulants (CVDP002AF)" "AF: Treated with anticoagulants (CVDP002AF)" ...
#>   ..$ NotificationCount     : int [1:22] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ TimePeriodID          : num [1:22] 17 17 17 17 17 17 17 17 17 17 ...
#>   ..$ TimePeriodName        : chr [1:22] "To March 2024" "To March 2024" "To March 2024" "To March 2024" ...
#>   ..$ CategoryAttribute     : chr [1:22] "Persons" "Female" "Male" "Male" ...
#>   ..$ MetricCategoryID      : int [1:22] 30 28 29 4 5 6 7 8 9 10 ...
#>   ..$ MetricCategoryName    : chr [1:22] "Persons" "Female" "Male" "18-39" ...
#>   ..$ MetricCategoryOrder   : int [1:22] 1 2 3 7 8 9 10 11 12 13 ...
#>   ..$ MetricCategoryTypeName: chr [1:22] "Sex" "Sex" "Sex" "Age group" ...
#>   ..$ MetricID              : int [1:22] 150 125 124 127 131 134 126 132 135 128 ...
#>  $ area_data        : tibble [22 × 27] (S3: tbl_df/tbl/data.frame)
#>   ..$ MetricID                       : int [1:22] 150 125 124 127 131 134 126 132 135 128 ...
#>   ..$ MetricCategoryTypeName         : chr [1:22] "Sex" "Sex" "Sex" "Age group" ...
#>   ..$ MetricCategoryName             : chr [1:22] "Persons" "Female" "Male" "18-39" ...
#>   ..$ CategoryAttribute              : chr [1:22] "Persons" "Female" "Male" "Male" ...
#>   ..$ AreaCode                       : chr [1:22] "U77315" "U77315" "U77315" NA ...
#>   ..$ AreaID                         : num [1:22] 701 701 701 NA NA NA 701 701 701 701 ...
#>   ..$ AreaName                       : chr [1:22] "Leicester Central PCN" "Leicester Central PCN" "Leicester Central PCN" NA ...
#>   ..$ Count                          : num [1:22] 1274 1274 1274 NA NA ...
#>   ..$ DataID                         : num [1:22] 7987516 7968114 7969045 NA NA ...
#>   ..$ Denominator                    : num [1:22] 175 75 100 NA NA NA 10 10 15 60 ...
#>   ..$ Factor                         : logi [1:22] NA NA NA NA NA NA ...
#>   ..$ HighestPriorityNotificationType: logi [1:22] NA NA NA NA NA NA ...
#>   ..$ LowerConfidenceLimit           : num [1:22] 91.5 87.2 90.2 NA NA ...
#>   ..$ Max                            : num [1:22] 97.2 98.6 100 NA NA ...
#>   ..$ Median                         : num [1:22] 91.7 91.5 92 NA NA ...
#>   ..$ Min                            : num [1:22] 54.8 52.4 56.5 NA NA ...
#>   ..$ NotificationCount              : int [1:22] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ Numerator                      : num [1:22] 165 70 95 NA NA NA 10 10 15 55 ...
#>   ..$ Q20                            : num [1:22] 54.8 52.4 56.5 NA NA ...
#>   ..$ Q40                            : num [1:22] 90.2 89.5 90.4 NA NA ...
#>   ..$ Q60                            : num [1:22] 91.7 91.5 92 NA NA ...
#>   ..$ Q80                            : num [1:22] 93.2 93.1 93.5 NA NA ...
#>   ..$ TimePeriodID                   : num [1:22] 17 17 17 NA NA NA 17 17 17 17 ...
#>   ..$ TimePeriodName                 : chr [1:22] "To March 2024" "To March 2024" "To March 2024" "To March 2024" ...
#>   ..$ UpperConfidenceLimit           : num [1:22] 97.7 97.4 98 NA NA ...
#>   ..$ Value                          : num [1:22] 95.4 94.7 96 NA NA ...
#>   ..$ ValueNote                      : chr [1:22] NA NA NA NA ...
#>  $ national_data    : tibble [22 × 26] (S3: tbl_df/tbl/data.frame)
#>   ..$ MetricID                       : int [1:22] 150 125 124 127 131 134 126 132 135 128 ...
#>   ..$ MetricCategoryTypeName         : chr [1:22] "Sex" "Sex" "Sex" "Age group" ...
#>   ..$ MetricCategoryName             : chr [1:22] "Persons" "Female" "Male" "18-39" ...
#>   ..$ CategoryAttribute              : chr [1:22] "Persons" "Female" "Male" "Male" ...
#>   ..$ AreaCode                       : chr [1:22] "E92000001" "E92000001" "E92000001" "E92000001" ...
#>   ..$ AreaID                         : int [1:22] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ AreaName                       : chr [1:22] "England" "England" "England" "England" ...
#>   ..$ Count                          : logi [1:22] NA NA NA NA NA NA ...
#>   ..$ DataID                         : int [1:22] 7962889 7962864 7962863 7962872 7962868 7962873 7962866 7962867 7962874 7962871 ...
#>   ..$ Denominator                    : num [1:22] 1012168 454503 557665 354 426 ...
#>   ..$ Factor                         : logi [1:22] NA NA NA NA NA NA ...
#>   ..$ HighestPriorityNotificationType: logi [1:22] NA NA NA NA NA NA ...
#>   ..$ LowerConfidenceLimit           : num [1:22] 91.4 91.1 91.7 69.2 57.8 ...
#>   ..$ Max                            : logi [1:22] NA NA NA NA NA NA ...
#>   ..$ Median                         : logi [1:22] NA NA NA NA NA NA ...
#>   ..$ Min                            : logi [1:22] NA NA NA NA NA NA ...
#>   ..$ NotificationCount              : int [1:22] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ Numerator                      : num [1:22] 925978 414349 511629 262 266 ...
#>   ..$ Q20                            : logi [1:22] NA NA NA NA NA NA ...
#>   ..$ Q40                            : logi [1:22] NA NA NA NA NA NA ...
#>   ..$ Q60                            : logi [1:22] NA NA NA NA NA NA ...
#>   ..$ Q80                            : logi [1:22] NA NA NA NA NA NA ...
#>   ..$ TimePeriodID                   : int [1:22] 17 17 17 17 17 17 17 17 17 17 ...
#>   ..$ UpperConfidenceLimit           : num [1:22] 91.5 91.2 91.8 78.2 66.9 ...
#>   ..$ Value                          : num [1:22] 91.5 91.2 91.7 74 62.4 ...
#>   ..$ ValueNote                      : logi [1:22] NA NA NA NA NA NA ...