Time periods and system levels

Returns all available time periods along with the systems levels included in each time period.

Usage

cvd_time_period_system_levels()

Value

tibble of time periods and associated system levels

Details

CVD Prevent API documentation: Time period system levels

See also

Examples

# get a tibble of all periods and levels
periods_levels <- cvd_time_period_system_levels()

# see which levels are available for the latest period
periods_levels |>
  dplyr::filter(TimePeriodID == max(TimePeriodID)) |>
  dplyr::select(TimePeriodID, TimePeriodName, SystemLevelID, SystemLevelName)
#> # A tibble: 3 × 4
#>   TimePeriodID TimePeriodName      SystemLevelID SystemLevelName
#>          <int> <chr>                       <int> <chr>          
#> 1           21 Jul 2023 - Jun 2024             1 England        
#> 2           21 Jul 2023 - Jun 2024             6 Region         
#> 3           21 Jul 2023 - Jun 2024             7 ICB