Search for matching areas

Returns a list of Areas that match a partial name for a given time period. Uses simple LIKE '%<partial_area_name>%' comparison.

Usage

cvd_area_search(partial_area_name = "Surgery", time_period_id = 1)

Arguments

partial_area_name

string - string to use to search for an Area (compulsory)

time_period_id

integer - limits the search to Areas which have data in specified time period (compulsory)

Value

Tibble of matching areas

Details

CVD Prevent API documentation: Area search

See also

Examples

# search for areas matching the term 'practice'
cvd_area_search(partial_area_name = 'practice', time_period_id = 17) |>
  dplyr::select(AreaID, AreaName, AreaCode)
#> # A tibble: 1,447 × 3
#>    AreaID AreaName                             AreaCode
#>     <int> <chr>                                <chr>   
#>  1   7735 1/Monton Medical Practice            P87620  
#>  2   7788 1/SALFORD MEDICAL PRACTICE           P87004  
#>  3   6408 3/Lower Broughton Medical Practice   P87654  
#>  4   6396 3/Springfield House Medical Practice P87024  
#>  5   6794 ABBEY GRANGE MEDICAL PRACTICE        B86068  
#>  6   4804 Abbey House Medical Practice         K83032  
#>  7   4814 Abbey Medical Practice               K83047  
#>  8   2327 Abbey Medical Practice               C83051  
#>  9   5262 Abbey Medical Practice               M81094  
#> 10   3607 Abbey Road Medical Practice          F84111  
#> # ℹ 1,437 more rows

# search for areas matching the term 'PCN'
cvd_area_search(partial_area_name = 'PCN', time_period_id = 17) |>
  dplyr::select(AreaID, AreaName, AreaCode)
#> # A tibble: 1,272 × 3
#>    AreaID AreaName                  AreaCode
#>     <int> <chr>                     <chr>   
#>  1   1103 3 Centres PCN             U60176  
#>  2    920 4 Doncaster PCN           U72999  
#>  3   1066 4PCN (Bnssg) PCN          U16600  
#>  4    923 A1 Network PCN            U89554  
#>  5    640 A31 Group PCN             U72748  
#>  6   1038 A34 West Berkshire PCN    U16983  
#>  7    222 Abbey Health PCN          U06079  
#>  8    174 Abc PCN                   U47228  
#>  9    889 Abingdon And District PCN U57321  
#> 10   1074 Abingdon Central PCN      U15230  
#> # ℹ 1,262 more rows