List system levels per time period

Returns all available system levels for a specified time period.

Usage

cvd_area_system_level(time_period_id = 1)

Arguments

time_period_id

integer - the time period to return data for (compulsory)

Value

tibble of system levels available for the time period

Details

CVD Prevent API documentation: System levels per time period

See also

Examples

# list system levels for time period 4
cvd_area_system_level(time_period_id = 4) |>
  dplyr::select(SystemLevelID, SystemLevelName)
#> # A tibble: 6 × 2
#>   SystemLevelID SystemLevelName
#>           <int> <chr>          
#> 1             1 England        
#> 2             6 Region         
#> 3             2 STP            
#> 4             3 CCG            
#> 5             4 PCN            
#> 6             5 Practice