Indicator data

Returns all metric data for a specified indicator. Data will include values for both selected area, and organisation at National Level (usually England).

Usage

cvd_indicator_data(indicator_id = 2, time_period_id = 1, area_id = 2)

Arguments

indicator_id

integer - indicator for which to return data (compulsory)

time_period_id

integer - time period for which to return data for (compulsory)

area_id

integer - area for which to return data for (compulsory)

Value

tibble of indicator data

Details

CVD Prevent API documentation: Indicator data

See also

Examples

test <- cvd_indicator_data()