Indicator group

Returns a single indicator group for a given group ID. An error will be returned if there is no indicator group associated with the given ID. IndicatorGroup is the primary key in the IndicatorGroup table, which also contains IndicatorGroupName and IndicatorGroupTypeID. The group type ID tells you what type of indicator group you're dealing with, e.g. a Priority Group. IndicatorGroupTypeID is the primary key of IndicatorGroupType and so IndicatorGroupTypeName is the associated name for the given group type ID. Finally, there is the array of indicators which are contained in this group, including display orders for the given group.

Usage

cvd_indicator_group(indicator_group_id = 15)

Arguments

indicator_group_id

integer - the group to return data for (compulsory)

Value

Tibble of indicators grouped by indicator group

Details

CVD Prevent API documentation: Indicator group

See also

Examples

#  list the indicators under Indicator Group ID 13 (Monitoring) which lists
# 'Key Question' Indicator Group indicators:
cvd_indicator_group(indicator_group_id = 13) |>
  dplyr::select(IndicatorGroupID, IndicatorGroupName, IndicatorGroupTypeName,
  IndicatorID, IndicatorName)
#> # A tibble: 2 × 5
#>   IndicatorGroupID IndicatorGroupName IndicatorGroupTypeName IndicatorID
#>              <int> <chr>              <chr>                        <int>
#> 1               13 Monitoring         Key Question                     4
#> 2               13 Monitoring         Key Question                    29
#> # ℹ 1 more variable: IndicatorName <chr>