Metric data

Returns all metric data for a specified metric. Data will include values for both selected area and organisation at National Level (usually England).

Usage

cvd_indicator_metric_data(metric_id, time_period_id, area_id)

Arguments

metric_id

integer - metric for which to return data for (compulsory)

time_period_id

integer - time period for which to return data for (compulsory)

area_id

integer - area for which to return data for (compulsory)

Value

Tibble of details for metric performance in the area and time period

Details

CVD Prevent API documentation: Indicator metric data

See also

Examples

# Retrieve metric data for AF: treatment with anticoagulants for 'Alliance
# PCN' (area ID 399) in time period 1 for metric 126 (breakdown by age group:
# males aged 40-59):
cvd_indicator_metric_data(metric_id = 126, time_period_id = 1, area_id = 399) |>
  dplyr::glimpse()
#> List of 3
#>  $ metrics      : tibble [1 × 17] (S3: tbl_df/tbl/data.frame)
#>   ..$ AxisCharacter         : chr "%"
#>   ..$ FormatDisplayName     : chr "Proportion %"
#>   ..$ IndicatorCode         : chr "CVDP002AF"
#>   ..$ IndicatorFormatID     : int 1
#>   ..$ IndicatorID           : int 7
#>   ..$ IndicatorName         : chr "Patients with GP recorded atrial fibrillation and with a CHADS2 or CHA2DS2-VASc score of 2 or more, who are cur"| __truncated__
#>   ..$ IndicatorOrder        : int 2
#>   ..$ IndicatorShortName    : chr "AF: Treated with anticoagulants (CVDP002AF)"
#>   ..$ NotificationCount     : int 0
#>   ..$ TimePeriodID          : int 1
#>   ..$ TimePeriodName        : chr "To March 2020"
#>   ..$ CategoryAttribute     : chr "Male"
#>   ..$ MetricCategoryID      : int 7
#>   ..$ MetricCategoryName    : chr "40-59"
#>   ..$ MetricCategoryOrder   : int 10
#>   ..$ MetricCategoryTypeName: chr "Age group"
#>   ..$ MetricID              : int 126
#>  $ area_data    : tibble [1 × 26] (S3: tbl_df/tbl/data.frame)
#>   ..$ MetricID                       : int 126
#>   ..$ MetricCategoryTypeName         : chr "Age group"
#>   ..$ MetricCategoryName             : chr "40-59"
#>   ..$ CategoryAttribute              : chr "Male"
#>   ..$ AreaCode                       : chr "U75382"
#>   ..$ AreaID                         : int 399
#>   ..$ AreaName                       : chr "Alliance PCN"
#>   ..$ Count                          : num 1214
#>   ..$ DataID                         : int 150124
#>   ..$ Denominator                    : logi NA
#>   ..$ Factor                         : num 1
#>   ..$ HighestPriorityNotificationType: logi NA
#>   ..$ LowerConfidenceLimit           : logi NA
#>   ..$ Max                            : num 100
#>   ..$ Median                         : num 84.6
#>   ..$ Min                            : num 60
#>   ..$ NotificationCount              : int 0
#>   ..$ Numerator                      : logi NA
#>   ..$ Q20                            : num 60
#>   ..$ Q40                            : num 78.4
#>   ..$ Q60                            : num 84.6
#>   ..$ Q80                            : num 91.2
#>   ..$ TimePeriodID                   : int 1
#>   ..$ UpperConfidenceLimit           : logi NA
#>   ..$ Value                          : logi NA
#>   ..$ ValueNote                      : chr "Value suppressed for disclosure control due to small count"
#>  $ national_data: tibble [1 × 26] (S3: tbl_df/tbl/data.frame)
#>   ..$ MetricID                       : int 126
#>   ..$ MetricCategoryTypeName         : chr "Age group"
#>   ..$ MetricCategoryName             : chr "40-59"
#>   ..$ CategoryAttribute              : chr "Male"
#>   ..$ AreaCode                       : chr "E92000001"
#>   ..$ AreaID                         : int 1
#>   ..$ AreaName                       : chr "England"
#>   ..$ Count                          : logi NA
#>   ..$ DataID                         : int 126
#>   ..$ Denominator                    : num 14028
#>   ..$ Factor                         : num 1
#>   ..$ HighestPriorityNotificationType: logi NA
#>   ..$ LowerConfidenceLimit           : num 81.9
#>   ..$ Max                            : logi NA
#>   ..$ Median                         : logi NA
#>   ..$ Min                            : logi NA
#>   ..$ NotificationCount              : int 0
#>   ..$ Numerator                      : num 11576
#>   ..$ Q20                            : logi NA
#>   ..$ Q40                            : logi NA
#>   ..$ Q60                            : logi NA
#>   ..$ Q80                            : logi NA
#>   ..$ TimePeriodID                   : int 1
#>   ..$ UpperConfidenceLimit           : num 83.1
#>   ..$ Value                          : num 82.5
#>   ..$ ValueNote                      : chr ""