Indicator priority groups

Returns the list of top-level groupings (Priority Groups) displayed in the Regional & ICS Insights page. Returns a dictionary called 'PriorityGroups' with each key being a Priority Group name, and each value being the array of indicators contained in that group. The 'PriorityGroupDisplayOrder' indicates the order in which it should be displayed for the given Priority Group.

Usage

cvd_indicator_priority_groups()

Value

tibble of data

Details

CVD Prevent API documentation: Indicator priority groups

See also

Examples

# Return one indicator from each of the priority groups:
cvd_indicator_priority_groups() |>
  dplyr::select(PriorityGroup, PathwayGroupName, PathwayGroupID, IndicatorCode, IndicatorID, IndicatorName) |>
  dplyr::slice_head(by = PathwayGroupID)
#> # A tibble: 7 × 6
#>   PriorityGroup   PathwayGroupName      PathwayGroupID IndicatorCode IndicatorID
#>   <chr>           <chr>                          <dbl> <chr>               <int>
#> 1 ABC             Hypertension                       5 CVDP004HYP              4
#> 2 ABC             Cholesterol                        6 CVDP006CHOL            22
#> 3 ABC             Atrial Fibrilation                 7 CVDP002AF               7
#> 4 CKD             Chronic Kidney Disea…              9 CVDP006CKD             29
#> 5 Prevalence      NA                                NA CVDP001CVD             12
#> 6 Prevalence      Familial Hypercholes…              8 CVDP002FH               9
#> 7 Smoking and BMI Smoking                           10 CVDP001SMOK            24
#> # ℹ 1 more variable: IndicatorName <chr>